extract edited PIC codec into dedicated microcrate
