+++ Cargo.toml
@@ -21,6 +21,7 @@ members = [
     "crates/copybook-corruption-rdw",
     "crates/copybook-corruption-predicates",
     "crates/copybook-dialect",
+    "crates/copybook-edited-pic",
     "crates/copybook-codec",
     "crates/copybook-codec-memory",
     "crates/copybook-sequence-ring",
@@ -87,6 +88,7 @@ copybook-corruption-detectors = { version = "=0.4.3", path = "crates/copybook-co
 copybook-corruption-rdw = { version = "=0.4.3", path = "crates/copybook-corruption-rdw" }
 copybook-corruption-predicates = { version = "=0.4.3", path = "crates/copybook-corruption-predicates" }
 copybook-dialect = { version = "=0.4.3", path = "crates/copybook-dialect" }
+copybook-edited-pic = { version = "=0.4.3", path = "crates/copybook-edited-pic" }
 copybook-codec = { version = "=0.4.3", path = "crates/copybook-codec" }
 copybook-codec-memory = { version = "=0.4.3", path = "crates/copybook-codec-memory" }
 copybook-cli-determinism = { version = "=0.4.3", path = "crates/copybook-cli-determinism" }
