+++ crates/copybook-edited-pic/README.md
@@ -1,0 +1,3 @@
+# copybook-edited-pic
+
+Single-responsibility microcrate for edited COBOL PIC pattern tokenization and numeric decode/encode.
